Organize pulsera files into frontend and backend
