+++ pulseras/backend/pulsera_functions.php
