+++ pulseras/backend/register_pulser_manual.php
@@ -1,6 +1,6 @@
 <?php
 session_start();
-require_once '../config/config.php';
+require_once '../../config/config.php';
 
 if ($_SERVER['REQUEST_METHOD'] === 'POST') {
     $id_usuario = $_SESSION['user_id'];
