+++ pulseras/backend/validar_codigo_invitacion.php
@@ -1,6 +1,6 @@
 <?php
 session_start();
-require_once '../config/config.php';
+require_once '../../config/config.php';
 
 // Verificar si el usuario está logueado
 if (!isset($_SESSION['user_id'])) {
