+++ pulseras/frontend/templates/layout.php
@@ -7,7 +7,7 @@
     <!-- Bootstrap CSS -->
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet">
     <script src="https://code.jquery.com/jquery-3.6.0.min.js"></script>
-    <link href="../colors.css" rel="stylesheet">
+    <link href="../../colors.css" rel="stylesheet">
     <?php if (isset($additionalHeadContent)) echo $additionalHeadContent; ?>
 </head>
 <body class="">
